Modified ComposedAuthorizationPlugin to use a custom ClassFactory
